requirements.txt
flask
opencv-python
torch
torchvision
matplotlib
seaborn
pandas
pillow
numpy
pathlib
